Fixing configs for css modules
